Update previous patch for current codebase.
