--- core/jvm/src/test/scala/matryoshka/exp/package.scala
@@ -1,30 +1,0 @@
-/*
- * Copyright 2014–2016 SlamData Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package matryoshka
-
-import matryoshka.data.Fix
-
-import scala.{Int, Symbol}
-
-package object exp {
-  def num(v: Int) = Fix[Exp](Num(v))
-  def mul(left: Fix[Exp], right: Fix[Exp]) = Fix[Exp](Mul(left, right))
-  def vari(v: Symbol) = Fix[Exp](Var(v))
-  def lam(param: Symbol, body: Fix[Exp]) = Fix[Exp](Lambda(param, body))
-  def ap(func: Fix[Exp], arg: Fix[Exp]) = Fix[Exp](Apply(func, arg))
-  def let(name: Symbol, v: Fix[Exp], inBody: Fix[Exp]) = Fix[Exp](Let(name, v, inBody))
-}
