+++ tests/src/test/scala/matryoshka/helpers/AlgebraChecks.scala
