+++ tests/src/test/scala/matryoshka/instances/fixedpoint/stream.scala
