+++ tests/src/test/scala/matryoshka/patterns/coenv.scala
