+++ tests/src/test/scala/matryoshka/patterns/diff.scala
