+++ tests/src/test/scala/matryoshka/patterns/envt.scala
