+++ tests/src/test/scala/matryoshka/patterns/potentialFailure.scala
