+++ tests/src/test/scala/matryoshka/specs2/scalacheck/CheckAll.scala
